+++ src/views/partials/nav.blade.php
@@ -15,9 +15,14 @@
 
 		<div class="navbar-collapse collapse" id="bs-example-navbar-collapse-1">
 			<ul class="nav navbar-nav navbar-left">
-				<li class="active"><a href="{{URL::to('/admin')}}"><i class="fa fa-home"></i> Home</a></li>
+				<li class="{{ $menu->activeClassForUrl('admin') }}">
+					<a href="{{ URL::to('admin') }}">
+						<i class="fa fa-home"></i>
+						Home
+					</a>
+				</li>
 				@foreach($menu->getTopLevelItems() as $item => $items)
-					<li class="dropdown">
+					<li class="dropdown {{ $menu->activeClassForUrls($items) }}">
 						<a href="#" class="dropdown-toggle" data-toggle="dropdown">
 							<i class="fa fa-file-text"></i>
 							{{ $item }}
@@ -25,12 +30,17 @@
 						</a>
 						<ul class="dropdown-menu">
 							@foreach ($items as $item => $link)
-								<li>{{ link_to($link, $item) }}</li>
+								<li class="{{ $menu->activeClassForPartialUrl($link) }}">{{ link_to($link, $item) }}</li>
 							@endforeach
 						</ul>
 					</li>
 				@endforeach
-				<li><a href="{{URL::to('/admin/users')}}"><i class="fa fa-user"></i> Users</a></li>
+				<li class="{{ $menu->activeClassForPartialUrl('admin/users') }}">
+					<a href="{{URL::to('admin/users')}}">
+						<i class="fa fa-user"></i>
+						Users
+					</a>
+				</li>
 			</ul>
 			<ul class="nav navbar-nav navbar-right">
 				<li class="dropdown">